+++ eras/conway/impl/testlib/Test/Cardano/Ledger/Conway/TreeDiff.hs
@@ -126,8 +126,18 @@ instance
       (a, b) = finishDRepPulser x
 
 instance ToExpr (RatifyEnv era) where
-  toExpr (RatifyEnv stake pool drep dstate ep cs) =
-    App "RatifyEnv" [toExpr stake, toExpr pool, toExpr drep, toExpr dstate, toExpr ep, toExpr cs]
+  toExpr (RatifyEnv stake pool drep dstate ep cs delegatees poolps) =
+    App
+      "RatifyEnv"
+      [ toExpr stake
+      , toExpr pool
+      , toExpr drep
+      , toExpr dstate
+      , toExpr ep
+      , toExpr cs
+      , toExpr delegatees
+      , toExpr poolps
+      ]
 
 -- Rules/Gov
 instance (EraPParams era, ToExpr (PParamsHKD StrictMaybe era)) => ToExpr (ConwayGovPredFailure era)