+++ libs/cardano-ledger-conformance/src/Test/Cardano/Ledger/Conformance/ExecSpecRule/Conway/Base.hs
@@ -57,7 +57,9 @@ import Cardano.Ledger.Conway.Governance (
   RatifySignal (..),
   RatifyState (..),
   VotingProcedures,
+  ensProtVerL,
   gasAction,
+  rsEnactStateL,
   showGovActionType,
  )
 import Cardano.Ledger.Conway.Rules (
@@ -85,7 +87,7 @@ import qualified Data.Sequence.Strict as SSeq
 import qualified Data.Set as Set
 import qualified Data.Text as T
 import GHC.Generics (Generic)
-import Lens.Micro (Lens', lens, (&), (.~))
+import Lens.Micro (Lens', lens, (&), (.~), (^.))
 import qualified Lib as Agda
 import qualified Prettyprinter as PP
 import Test.Cardano.Ledger.Binary.TreeDiff (tableDoc)
@@ -315,7 +317,7 @@ ratifyEnvSpec ::
   ConwayRatifyExecContext Conway ->
   Specification fn (RatifyEnv Conway)
 ratifyEnvSpec ConwayRatifyExecContext {crecGovActionMap} =
-  constrained' $ \_ poolDistr drepDistr drepState _ committeeState ->
+  constrained' $ \_ poolDistr drepDistr drepState _ committeeState _ _ ->
     [ -- Bias the generator towards generating DReps that have stake and are registered
       exists
         ( \eval ->
@@ -485,6 +487,7 @@ instance IsConwayUniv fn => ExecSpecRule fn "RATIFY" Conway where
                 <*> toSpecRep st
                 <*> toSpecRep sig
           ]
+      pv = st ^. rsEnactStateL . ensProtVerL
       actionAcceptedRatio gas@GovActionState {..} =
         tableDoc
           (Just "GovActionState")
@@ -495,7 +498,7 @@ instance IsConwayUniv fn => ExecSpecRule fn "RATIFY" Conway where
           ,
             ( "SPO:"
             , showAccepted (spoAccepted env st gas)
-                PP.<+> showRatio (spoAcceptedRatio env gas)
+                PP.<+> showRatio (spoAcceptedRatio env gas pv)
             )
           ,
             ( "DRep:"
@@ -524,6 +527,7 @@ instance IsConwayUniv fn => ExecSpecRule fn "RATIFY" Conway where
         | otherwise = error "ratio is not in the unit interval"
       committee = ensCommittee rsEnactState
       members = foldMap' (committeeMembers @Conway) committee
+      pv = st ^. rsEnactStateL . ensProtVerL
       ccBucket a =
         "CC yes votes ratio  \t"
           <> bucket
@@ -540,7 +544,7 @@ instance IsConwayUniv fn => ExecSpecRule fn "RATIFY" Conway where
       spoBucket a =
         "SPO yes votes ratio \t"
           <> bucket
-            (spoAcceptedRatio env a)
+            (spoAcceptedRatio env a pv)
       acceptedActions = fmap gasAction . filter (acceptedByEveryone env st) $ toList actions
       acceptedActionTypes = Set.fromList $ showGovActionType <$> acceptedActions
       labelRatios